edit tf.constant

+++ model.py
@@ -154,7 +154,7 @@ class ClaimBusterLayer(K.layers.Layer):
             y_int = tf.argmax(y, axis=1)
             print(y_int)
             # loss_adj = [self.computed_cls_weights[x] for x in y_int]
-            cw = tf.constant(self.computed_cls_weights)
+            cw = tf.constant(self.computed_cls_weights, dtype=tf.float32)
             print(cw)
             loss_adj = tf.map_fn(lambda x: cw, y_int)
             print(loss_adj)
